Changing type

+++ data/request/updateRentalRequest.go
@@ -6,7 +6,7 @@ type UpdateRentalRequest struct {
 	LocationID  uint    `json: validate:"required"`
 	AccountID   uint    `json: validate:"required"`
 	Bedrooms    uint    `json: validate:"required"`
-	Bathrooms   uint    `json: validate:"required"`
+	Bathrooms   float64 `json: validate:"required"`
 	NightlyRate float64 `json: validate:"required"`
 	Description string  `json: validate:"required"`
 }
